requirements.txt
aiogram==3.7.0
gspread==6.0.2
google-auth==2.28.0
python-dotenv==1.0.1